fix: 优化增加赛季时，如果没有开启的赛季，新增时，进行开启，心愿增加配置兑换限制

+++ src/main/java/com/tencent/wxcloudrun/dao/SeasonConfigMapper.java
@@ -41,7 +41,7 @@ public interface SeasonConfigMapper {
      * @param seasonConfig 赛季配置
      * @return 影响的行数
      */
-    int insert(SeasonConfig seasonConfig);
+    Long insert(SeasonConfig seasonConfig);
     
     /**
      * 更新赛季
